requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
coverage==5.5
Django==3.2.6
djangorestframework==3.12.4
gunicorn==20.0.4
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
docker
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
